requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
click
python-secrets
pycrypto
gunicorn
